consolidate thread mutations
