+++ convex/_generated/api.d.ts
@@ -21,7 +21,6 @@ import type * as db_imageModels from "../db/imageModels.js";
 import type * as db_images from "../db/images.js";
 import type * as db_messages from "../db/messages.js";
 import type * as db_models from "../db/models.js";
-import type * as db_runcreate from "../db/runcreate.js";
 import type * as db_speech from "../db/speech.js";
 import type * as db_threads from "../db/threads.js";
 import type * as db_voiceModels from "../db/voiceModels.js";
@@ -77,7 +76,6 @@ declare const fullApi: ApiFromModules<{
   "db/images": typeof db_images;
   "db/messages": typeof db_messages;
   "db/models": typeof db_models;
-  "db/runcreate": typeof db_runcreate;
   "db/speech": typeof db_speech;
   "db/threads": typeof db_threads;
   "db/voiceModels": typeof db_voiceModels;
